hf-spaces/image-captioner/requirements.txt
transformers
fastapi
uvicorn[standard]
Pillow
python-multipart
